chore(tags): extracted component to scenario
